pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.0.1"
description = ""
authors = ["Aaron Mamparo <aaronmamparo@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11,<3.14"
requests = "^2.32.3"
rss-parser = "^2.1.0"
atproto = "^0.0.56"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
boto3 = "^1.35.86"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
pylint-quotes = "^0.2.3"
aws-cdk-lib = "^2.173.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
